chore(lint): fix clippy lints where possible

+++ src/test/storage.rs
@@ -33,11 +33,11 @@ fn double_map_iterator_test() {
 	TestExternalities::default().execute_with(|| {
 		// shard 0
 		for i in 0..4 {
-			DoubleMap::insert(0 as u16, i as u32, i as u64);
+			DoubleMap::insert(0_u16, i as u32, i as u64);
 		}
 		// shard 1
 		for i in 0..5 {
-			DoubleMap::insert(1 as u16, i as u32, i as u64)
+			DoubleMap::insert(1_u16, i as u32, i as u64)
 		}
 
 		assert_eq!(
@@ -56,12 +56,12 @@ fn double_map_iterator_test() {
 		);
 
 		assert_eq!(
-			DoubleMap::iter_prefix(0 as u16).collect::<Vec<_>>(),
+			DoubleMap::iter_prefix(0_u16).collect::<Vec<_>>(),
 			vec![(0, 0), (1, 1), (2, 2), (3, 3)]
 		);
 
 		assert_eq!(
-			DoubleMap::iter_prefix_values(1 as u16).collect::<Vec<_>>(),
+			DoubleMap::iter_prefix_values(1_u16).collect::<Vec<_>>(),
 			vec![0, 1, 2, 3, 4]
 		);
 		let starting_raw_key = DoubleMap::hashed_key_for(1, 2);
